Revamped get/set Content.
Now you can get and set content using line, column or cellId.
Fixed permission conditional.
Throws a PermissionException if the cell is protected when you
setContent.
The reason I simplified AssignLiteralCell was one of the conditions to
write was present in setContent. But not if the Cell is protected. Since
it's a business i've added to setContent. Also verification if the cell
exists are done in CalcSheet, whether is with line;column or cellId.