Removed now unecessary code in ExportDocumentTest
Before there was an Integration layer this code was useful, since the
service had as a result the exported xml. Now the integration doesn't
return a result, it just sends the exported data to another service.
This test is supposed to test a service, not exporting XML, which should
be done in another test case.