SD-ID WORKS
Holy Shit! After goose chasing and git bissecting I discovered what was
making SD-ID call an abstract method, thus throwing AbstractMethodError.
IT WAS FUCKING JAXEN! Well, my hypothesis is that the JAX-WS uses jaxen,
but it's a different version of jaxen, and the new one covered it. Well,
that's only a hypothesis.